move legacy Tabs
