+++ src/components/AssetSelectorModal/AssetSelectorModal.js
@@ -27,7 +27,7 @@ import t from 'translations/translate';
 import ContainerWithHeader from 'components/legacy/Layout/ContainerWithHeader';
 import SearchBar from 'components/SearchBar';
 import SlideModal from 'components/Modals/SlideModal';
-import Tabs from 'components/Tabs';
+import Tabs from 'components/legacy/Tabs';
 
 // Constants
 import { ASSET_TYPES } from 'constants/assetsConstants';
