+++ src/components/legacy/ActivityFeed/ActivityFeed.js
@@ -29,7 +29,7 @@ import type { EnsRegistry } from 'reducers/ensRegistryReducer';
 
 // components
 import Title from 'components/Title';
-import Tabs from 'components/Tabs';
+import Tabs from 'components/legacy/Tabs';
 import { BaseText, MediumText } from 'components/legacy/Typography';
 import EmptyStateParagraph from 'components/EmptyState/EmptyStateParagraph';
 import ActivityFeedItem from 'components/legacy/ActivityFeed/ActivityFeedItem';
