+++ src/components/legacy/Tabs/package.json
